LayoutTests/media/media-source/media-source-canplaythrough-event-expected.txt


EVENT(sourceopen)
EVENT(canplaythrough)
END OF TEST
